Remove os-release diversions once rootfs creation is finished
This is necessary to avoid overwrites while packages are being installed
when the chroot is being built up. Once it's done, it's no longer necessary.

This fixes a problem where Calamares rsyncs the whole read-only live rootfs
into an installed system, which then is supposed to get updated, and the
divert stops the file from being updated as expected, which was never
intended.

Follow-up for 24bfe093f82ccc134ab5d91fb4058ee0b52ec942

Partially addresses #1111039