installer: Allow the grub signed packages to be missing, don't fail
Some Debian derivatives don't have the signed GRUB packages. For
example, Kali Linux has its own fork of GRUB (it doesn't use it from
Debian "as is"), but has no way to sign it, so the packages
grub-efi-$ARCH-signed simply don't exist in Kali.

With this commit, live-build handles this situation gracefully: it
doesn't require these packages if they are not available.